datapeace/requirements.txt
bson==0.5.6
Django==2.0.7
djangorestframework==3.8.2
pip==19.1.1
psycopg2==2.7.5
psycopg2-binary==2.8.2
python-dateutil==2.8.0
pytz==2019.1
setuptools==41.0.1
six==1.12.0
wheel==0.33.4
